*** empty log message ***

+++ trunk/tests/Process_Strategy_Test.cpp
@@ -388,7 +388,7 @@ Counting_Service::handle_input (ACE_HANDLE)
 
   bytes = this->peer ().recv (buf, len);
 
-  if (bytes <= 0 || buf[0] == EOF)
+  if (bytes <= 0 || (int) buf[0] == EOF)
     return -1;
   else
     {
